[ALPS07927432] hif: Fix coverity issue
Invalid type in argument to printf format specifier: 10641290
Out-of-bounds read: 11362120

MTK-Commit-Id: e858a8fc0d7430c9e56b9d49badcec09c8712464

CR-Id: ALPS07927432
Feature: Wi-Fi Driver CONNAC
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/7341940
CheckSWIP: srv_pis <srv_pis@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
Test: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Coverity-Review-Label: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_preflight_b001 <srv_preflight_b001@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
CheckCourse: srv_pis <srv_pis@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_pf_nep_sanity

+++ os/linux/hif/pcie/pcie.c
@@ -1795,9 +1795,9 @@ static void glBusFreeMsiIrq(struct pci_dev *pdev,
 
 		written += kalSnprintf(dbg + written,
 				       sizeof(dbg) - written,
-				       "[%d] %llu, ",
+				       "[%d] %u, ",
 				       irqn,
-				       KAL_GET_TIME_INTERVAL());
+				       (uint32_t)KAL_GET_TIME_INTERVAL());
 	}
 	KAL_REC_TIME_END();
